Disable event annotation option if visualization does not support it. (#10427)
This change is introducing a `capabilities` field for a visualization type, which can be used to describe which optional features a given visualization supports. The valid keys for this field are implemented through an interface, so it can be extended from plugins through declaration merging.

These capabilities are effectively used in this PR to disable event annotations for heatmap/single number/pie chart/world map visualizations, which do not support displaying event annotations.