Disabling Github Actions if PR is not from fork (#10430)
This change is temporarily disabling the Github Actions build if the current PR
is not from a fork. For these, we have a Jenkins run. For forks, we do not run
that one, so we should at least have a Github Actions build.

The reason for disabling Github Actions is that we are seeing test failures for
no reason repeatedly. It looks like there is some kind of resource exhaustion
that leads to not being able to bind a port reliably. We do not see this for
our Jenkins builds.